requirements.txt
psycopg2==2.7.1
Django==1.11.7
django-cors-headers==1.3.1
django-rest-auth==0.9.1
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
django-rest-swagger==2.1.2
dynamic-rest==1.5.0
Pillow==4.0.0
celery==4.1.0
Keras==2.1.3
tensorflow==1.1.0
requests==2.18.4
h5py==2.7.1
termcolor==1.1.0
django-heroku==0.2.0
gunicorn==19.7.1
djangorestframework-bulk==0.2.1
djangorestframework-filters==0.10.2
graphviz==0.8.2
elasticsearch-dsl==6.1.0
